requirements.txt
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
binaryornot==0.4.4
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.2.3
coverage==7.2.7
Django==4.2.3
django-browser-reload==1.11.0
django-debug-toolbar==4.2.0
django-tailwind==3.6.0
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.2
Faker==19.2.0
flake8==6.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
parameterized==0.9.0
Pillow==10.0.0
pluggy==1.2.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PySocks==1.7.1
pytest==7.4.0
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
watchdog==3.0.0
wsproto==1.2.0
